refactor(request): extract pull request repo info helpers
